Refactor game, player, scenery, metrics, shapes, and colors into modules
