+++ vegetable-junction/src/colors.rs
@@ -1,0 +1,7 @@
+pub type Colour = [f32; 4];
+
+pub const RED: Colour = [1.0, 0.0, 0.0, 1.0];
+pub const GREEN: Colour = [0.0, 1.0, 0.0, 1.0];
+pub const BLUE: Colour = [0.0, 0.0, 1.0, 1.0];
+pub const WHITE: Colour = [1.0; 4];
+pub const BLACK: Colour = [0.0, 0.0, 0.0, 1.0];
