+++ vegetable-junction/src/metrics.rs
@@ -1,0 +1,8 @@
+#[derive(Default)]
+pub struct Metrics {
+    pub width: f64,
+    pub height: f64,
+}
+
+pub const WINDOW_WIDTH: f64 = 1500.0;
+pub const WINDOW_HEIGHT: f64 = 900.0;
